refactor: 경로 미리보기 화면 파일 분리 및 코드 정리
- RoutePreviewScreen.kt를 4개 파일로 분리 (729 -> 59 lines)
- RouteExtensions.kt: Route 및 List<LatLng> 헬퍼 확장 함수
- RouteMapView.kt: GoogleMap 렌더링 및 상단 칩 컴포넌트
- RouteDetailSheet.kt: 바텀시트 상세 내용 및 관련 컴포넌트
- 불필요한 코드 제거 및 컴포저블 함수 분리
